chore: refactored dungeon
