+++ src/model/dungeon/constants.ts
@@ -1,0 +1,10 @@
+import { RoomSizeRange } from './types';
+
+export const ROOM_SIZE_SMALL: RoomSizeRange = [45, 50];
+export const ROOM_SIZE_MEDIUM: RoomSizeRange = [50, 80];
+export const ROOM_SIZE_LARGE: RoomSizeRange = [80, 120];
+
+export const NUM_ROOMS_PER_CLICK = 3;
+export const CANVAS_SIZE = 1024;
+export const DOOR_WIDTH = 8;
+export const DOOR_HEIGHT = 12;
